feat(main): add token window and remove old windows
