--- src/main/app_windows/index.ts
@@ -1,9 +1,0 @@
-export { CreateAboutWindow } from './about_window';
-export { CreateKeyPinWindow } from './key_pin_window';
-export { CreateSettingsWindow } from './settings_window';
-export { CreateP11PinWindow } from './p11_pin_window';
-export { CreateErrorWindow } from './error_window';
-export { DialogsStorage } from './utils';
-export { IBrowserWindow } from './window';
-export { CreateWarningWindow } from './warning_window';
-export { CreateQuestionWindow } from './question_window';
