+++ src/main/main.ts
@@ -8,7 +8,6 @@
 import {
   app,
   ipcMain,
-  screen,
   shell,
   IpcMainEvent,
 } from 'electron';
@@ -180,11 +179,17 @@ async function InitService() {
       const atr = card.atr.toString('hex');
       winston.info(`New token was found reader: '${card.reader}' ATR: ${atr}`);
 
-      CreateTokenWindow(
-        intl('question.new.token'),
-        { id: 'question.new.token', showAgain: true },
-        (res) => {
-          if (res) {
+      CreateTokenWindow({
+        params: {
+          type: 'token',
+          text: intl('question.new.token'),
+          id: 'question.new.token',
+          showAgain: true,
+          showAgainValue: false,
+          result: 0,
+        },
+        onClosed: (result) => {
+          if (result) {
             try {
               const title = `Add support for '${atr}' token`;
               const body = fs.readFileSync(TEMPLATE_NEW_CARD_FILE, { encoding: 'utf8' })
@@ -201,7 +206,30 @@ async function InitService() {
             }
           }
         },
-      );
+      });
+
+      // CreateTokenWindow(
+      //   intl('question.new.token'),
+      //   { id: 'question.new.token', showAgain: true },
+      //   (res) => {
+      //     if (res) {
+      //       try {
+      //         const title = `Add support for '${atr}' token`;
+      //         const body = fs.readFileSync(TEMPLATE_NEW_CARD_FILE, { encoding: 'utf8' })
+      //           .replace(/\$\{reader\}/g, card.reader)
+      //           .replace(/\$\{atr\}/g, atr.toUpperCase())
+      //           .replace(/\$\{driver\}/g, crypto.randomBytes(20).toString('hex').toUpperCase());
+      //         const url1 = `${SUPPORT_NEW_TOKEN_LINK}/issues/new?${querystring.stringify({
+      //           title,
+      //           body,
+      //         })}`;
+      //         shell.openExternal(url1);
+      //       } catch (e) {
+      //         winston.error(e.message);
+      //       }
+      //     }
+      //   },
+      // );
     })
     .on('error', (e: Error) => {
       winston.error(e.stack || e.toString());
