+++ src/main/windows/index.ts
@@ -1,18 +1,11 @@
-export { CreateMainWindow } from './main';
-export {
-  CreateTokenWindow,
-} from './message';
-export {
-  CreateWindow,
-  BrowserWindowEx,
-  BrowserWindowConstructorOptionsEx,
-} from './window';
-export {
-  CreateAboutWindow,
-  CreateKeyPinWindow,
-  CreateSettingsWindow,
-  CreateP11PinWindow,
-  CreateErrorWindow,
-  CreateWarningWindow,
-  CreateQuestionWindow,
-} from '../app_windows';
+export { CreateAboutWindow } from './about_window';
+export { CreateKeyPinWindow } from './key_pin_window';
+export { CreateSettingsWindow } from './settings_window';
+export { CreateP11PinWindow } from './p11_pin_window';
+export { CreateErrorWindow } from './error_window';
+export { DialogsStorage } from './utils';
+export { IBrowserWindow, BrowserWindow } from './window';
+export { CreateWarningWindow } from './warning_window';
+export { CreateQuestionWindow } from './question_window';
+export { CreateTokenWindow } from './token_window';
+export { CreateMainWindow } from './main_window';
