+++ src/main/windows/utils.ts
