+++ src/main/windows/window.ts
@@ -1,104 +1,139 @@
-import { app, BrowserWindow, shell, globalShortcut } from 'electron';
+/* eslint-disable class-methods-use-this */
+
+import {
+  BrowserWindow as ElectronWindow,
+  shell,
+  globalShortcut,
+} from 'electron';
 import * as url from 'url';
 import * as winston from 'winston';
 
 import { HTML_PATH, windowSizes, icons } from '../const';
 import { locale } from '../locale';
 
-let counter = 0;
-const windows: { [key: number]: BrowserWindowEx } = {};
+const isDev = process.env.NODE_ENV === 'development';
 
-export interface BrowserWindowEx extends Electron.BrowserWindow {
-  [key: string]: any;
+export interface IBrowserWindow extends ElectronWindow {
+  app: string;
   lang: string;
   params: Assoc<any>;
 }
 
-export interface BrowserWindowConstructorOptionsEx
-  extends Electron.BrowserWindowConstructorOptions {
+export interface IWindowOptions {
   app: string;
+  title: string;
+  size?: keyof typeof windowSizes;
   params?: Assoc<any>;
-  dock?: boolean;
+  onClosed: (...args: any[]) => void;
+  windowOptions?: {
+    modal?: boolean;
+    alwaysOnTop?: boolean;
+    x?: number;
+    y?: number;
+    center?: boolean;
+    parent?: ElectronWindow ;
+    show?: boolean;
+  };
 }
 
-export function CreateWindow(options: BrowserWindowConstructorOptionsEx) {
-  const isDev = process.env.NODE_ENV === 'development';
+export class BrowserWindow {
+  window: IBrowserWindow;
 
-  const window = new BrowserWindow({
-    icon: icons.favicon,
-    autoHideMenuBar: true,
-    dock: true,
-    minimizable: false,
-    fullscreen: false,
-    fullscreenable: false,
-    // Prevent resize window on production
-    resizable: isDev,
-    minWidth: windowSizes.small.width,
-    minHeight: windowSizes.small.height,
-    show: false,
-    ...options,
-    webPreferences: {
-      nodeIntegration: true,
-      // Prevent open DevTools on production
-      devTools: isDev,
-    },
-  }) as BrowserWindowEx;
+  constructor(options: IWindowOptions) {
+    this.window = new ElectronWindow({
+      title: options.title,
+      ...this.getWindowDefaultOptions(),
+      ...this.getWindowSize(options.size),
+      ...options.windowOptions,
+    }) as IBrowserWindow;
 
-  winston.info(`Fortify: Create window ${options.app}`);
+    this.onInit(options);
+  }
 
-  window.loadURL(url.format({
-    pathname: HTML_PATH,
-    protocol: 'file:',
-    slashes: true,
-  }));
+  private onInit(options: IWindowOptions) {
+    winston.info(`Fortify: Create window ${options.app}`);
 
-  window.lang = locale.lang;
-  window.app = options.app;
-  window.params = options.params || {};
+    this.window.loadURL(url.format({
+      pathname: HTML_PATH,
+      protocol: 'file:',
+      slashes: true,
+    }));
 
-  if ('dock' in app && options.dock) {
-    app.dock.show();
+    this.window.lang = locale.lang;
+    this.window.app = options.app;
+    this.window.params = options.params || {};
 
-    // Add window to list
-    const id = counter++;
-    windows[id] = window;
+    this.onInitListeners(options);
+  }
 
-    window.on('close', () => {
-      delete windows[id];
-
-      if (!Object.keys(windows).length) {
-        app.dock.hide();
+  private onInitListeners(options: IWindowOptions) {
+    // Open a url from <a> on default OS browser
+    this.window.webContents.on('will-navigate', (e: Event, href: string) => {
+      if (href !== this.window.webContents.getURL()) {
+        e.preventDefault();
+        shell.openExternal(href);
       }
     });
-  }
 
-  // Open a url from <a> on default OS browser
-  window.webContents.on('will-navigate', (e: Event, href: string) => {
-    if (href !== window.webContents.getURL()) {
-      e.preventDefault();
-      shell.openExternal(href);
+    // Show page only after `lfinish-load` event and prevent show index page
+    if (this.window.app !== 'index') {
+      this.window.webContents.on('did-finish-load', () => {
+        this.window.show();
+      });
     }
-  });
 
-  // Show page only after `lfinish-load` event and prevent show index page
-  if (options.app !== 'index') {
-    window.webContents.on('did-finish-load', () => {
-      window.show();
+    // Prevent BrowserWindow refreshes
+    this.window.on('focus', () => {
+      globalShortcut.registerAll(['CommandOrControl+R', 'F5'], () => {});
     });
+
+    this.window.on('blur', () => {
+      globalShortcut.unregisterAll();
+    });
+
+    this.window.on('close', () => {
+      globalShortcut.unregisterAll();
+    });
+
+    this.window.on('closed', options.onClosed);
   }
 
-  // Prevent BrowserWindow refreshes
-  window.on('focus', () => {
-    globalShortcut.registerAll(['CommandOrControl+R', 'F5'], () => {});
-  });
+  private getWindowDefaultOptions(): Electron.BrowserWindowConstructorOptions {
+    return {
+      icon: icons.favicon,
+      autoHideMenuBar: true,
+      minimizable: false,
+      fullscreen: false,
+      fullscreenable: false,
+      // Prevent resize window on production
+      resizable: isDev,
+      show: false,
+      ...this.getWindowSize(),
+      webPreferences: {
+        nodeIntegration: true,
+        // Prevent open DevTools on production
+        devTools: isDev,
+      },
+    };
+  }
 
-  window.on('blur', () => {
-    globalShortcut.unregisterAll();
-  });
+  private getWindowSize(size: keyof typeof windowSizes = 'default') {
+    if (size === 'small') {
+      return windowSizes.small;
+    }
 
-  window.on('close', () => {
-    globalShortcut.unregisterAll();
-  });
+    return windowSizes.default;
+  }
 
-  return window;
+  public focus() {
+    this.window.focus();
+  }
+
+  public show() {
+    this.window.show();
+  }
+
+  public hide() {
+    this.window.hide();
+  }
 }
